Add files via upload
The file is a reinforcing addition game that is used for the selection of miners in the blockchain system. In the reinforcing addition game, a  PRNG is employed to generate the necessary score values for the game. The initial seed is generated randomly by the proposed model.  A pseudo-random number is rolled 9 times and on each roll, one player displays the number, which is broadcasted to all players in the blockchain network